1.fix: Redis时间格式序列化错误, 不应该使用yyyy-MM-dd: HH:mm:ss格式, 导致反序列化时没有毫秒 2.fix: 其他一些错误

+++ hami-common/src/main/java/top/wang3/hami/common/dto/article/HotArticleDTO.java
@@ -11,6 +11,6 @@ import lombok.NoArgsConstructor;
 public class HotArticleDTO {
 
     private Integer articleId;
-    private Long hotRank;
+    private Double hotRank;
     private ArticleDTO article;
 }
